Photosurface example: scale and rotate the frame, not the image
The size of the border around the image could otherwise change
dramatically, because the PinchArea was zooming the image.
It's conceivable that the frame could follow the image in scale and
rotation, but that doesn't work as long as the image is inside the
frame, because changing scale and rotation of the parent affects the
child too.  So it follows that the scale and rotation of the image
should never be directly set.  This helps the apparent size of the
white border around each image to be visually consistent.  Another
advantage is that the the image centroid doesn't tend to move,
so compensation is not necessary.

Change-Id: Ib7767e369097b32f006274add73ac544aedaf839
Reviewed-by: Topi Reiniö <topi.reinio@digia.com>
Reviewed-by: Shawn Rutledge <shawn.rutledge@digia.com>