Fix GC crash with conditional breakpoints and JS console
We may choose to execute an expression in a specific frame within the
debugger, which is where we pop context's until we reached the frame
in question. If we are trying to execute an expression at the top of the
stack (or with a conditional breakpoint expression), then we don't have
a frame and don't need to pop contexts. But also also don't need to call
Scope::alloc(-1).

Change-Id: I1f6754a3d77d943aed9bf4468e817a5269a3c547
Reviewed-by: Lars Knoll <lars.knoll@digia.com>