QQuickWindow: prerequisites for handling multi-point touch
In touchEventForItem, we want to include only touchpoints
that occur inside the target item if the touchpoint is freshly
pressed; but if the touchpoint is being updated or released,
in most cases we want to keep sending it to the same item,
because the item must be able to change state.  But in
sendFilteredTouchEvent, we need to check the bounds of all
touchpoints.

Making touchEventForItem and touchEventWithPoints static
also avoids needing to get the QQuickWindowPrivate instance,
since these functions do not use any private data.  And we will
need a version of dragOverThreshold which works with TouchPoints
since it doesn't share a common ancestor class with QMouseEvent.

Change-Id: I022c6bcf36382741b31baeb020d3104d502a11e8
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>