Avoid assert in rich text when img width is invalid
If you set the width or height of an <img> tag to something
invalid, this will be registered as -2 by the HTML parser.
We should treat this case the same as if there is no width/height
specified and use the implicit size instead.

[ChangeLog][Text] Fixed assert when setting an invalid width or
height on an <img> tag in a text element.

Change-Id: Iae8c33fa184316632f72318e71f26ab005645a21
Task-number: QTBUG-44743
Reviewed-by: Gunnar Sletta <gunnar@sletta.org>