Fix flickable stealing gestures when height >= contentHeight.
Or width >= contentWidth.

If newY is equal to both maxY and minY then the second assignment to
rejectY can clobber the first.  Changing the second assignment to an
|= means rejectY is true if either is true.

Secondly maxY became positive when height was greater than contentHeight
instead of having a maximum of 0, which ensured rejectY evaluated to false
in the case newY <= maxY and the mouse was stolen.

Change-Id: I6416d0e23c3ef898887a7b3e3fcdc1dc12853548
Reviewed-by: Martin Jones <martin.jones@qinetic.com.au>