qmltest: Make sure we inherit all of testlib's MODULE_CONFIG
We don't know what's in testlib's MODULE_CONFIG, so we can't hard-code
the value. E.g., with the introduction of XCTest support in testlib [1],
we'll gain an additional member in testlib's module config set for taking
care of the XCTest framework include path, and this value also needs to
be propagated/inherited to the final target.

[1] https://codereview.qt-project.org/#/c/107477/

Change-Id: I28ee44ac41f148a873ab35adcae38e73bef3316f
Reviewed-by: Simon Hausmann <simon.hausmann@theqtcompany.com>