Android: Work around bug on Samsung Galaxy Tab 3 10.1
After installing the Android 4.4.2 upgrade on Samsung Galaxy Tab 3 10.1,
one of the vertex shaders in Qt Quick stopped compiling.

The issue seems to be that this particular shader triggers a
bug in the driver, and a simple work-around is to replace the
ternary conditional operator with a simple if-statement.

[ChangeLog][Android] Fixed a vertex shader compilation issue on
certain OpenGL drivers.

Change-Id: I2f878215a753a7e222c46c0f85a84b0bc81d523a
Task-number: QTBUG-43515
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>