Return the correct type from Item::mapToItem/Item::mapFromItem.
Previously we were returning a JavaScript object with x/y/width/height
properties, instead of a point/rect. This meant that the type couldn't
be converted to a point/rect because we don't support duck typing,
where we would deduce the type based on the properties.

One example of a broken use case that this patch fixes is when QML is
unable to convert the return type to a point in a property declaration:

property point p: mouseArea.mapToItem(child, mouseArea.mouseX, mouseArea.mouseY)

Another is using the result of the function to pass to another function:

child.contains(mouseArea.mapToItem(child, mouseArea.mouseX, mouseArea.mouseY))

Change-Id: I3ce82f10175f904dd02c8af6b5e42cee14b2ebb2
Task-number: QTBUG-41452
Reviewed-by: Liang Qi <liang.qi@theqtcompany.com>
Reviewed-by: Simon Hausmann <simon.hausmann@theqtcompany.com>