Speed up object creation with lots of id name objects
In the common case when objects are referenced by id, we don't need the
QQmlContextData's property name cache for the lookup. Instead of bindings
we resolve the id to an integer at type compilation time. Therefore we can
delay the population of the QQmlContextData's propertyNames until we need it
and therefore safe time on startup.

Change-Id: I2e8eeb51b1ac92f230059be3467f9a1e00fdbbef
Reviewed-by: Lars Knoll <lars.knoll@digia.com>