Make the scenegraph_lancelot test work on Mac
Grabbing the window after every frame is needlessly expensive and
would overload a less powerful Mac. Rewrote the grabber to just do the
necessary grabs.

Make the hostinfo script not output the unneeded '[undefined]' value
for unset environment variables, as the braces would upset the paths
in the html report.

Change-Id: I76104235c15eafd5e7f92f2ec01c84f5b11b55a6
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>