Stabilize tst_qquickimage::noLoading
The test loads an image from a remote source and it has a QSignalSpy installed
for the progressChanged(qreal) signal on the QQuickImage. The test counts the
number of times the signal is emitted and expects it to hit an exact value.
That seems wrong when loading an image over the network, as we have no control
over in what pieces the data arrives on our end. And as it turns out, the test
occasionally fails in the CI system, which is also reproducible.

To stabilize the test, we now expect at least two emissions of the signal when
loading from the network (0% and 100%) but there may be more (0% 10% 80% 100%
for example). In addition the reload case is simplified by resetting the spy
and verifying that progressChanged was not emitted at all.

Change-Id: Ib2e660651d40b92eff889ebe3baabb74d6e00fb4
Reviewed-by: Liang Qi <liang.qi@theqtcompany.com>
Reviewed-by: Albert Astals Cid <albert.astals@canonical.com>
Reviewed-by: Gunnar Sletta <gunnar@sletta.org>